Extract structs to lib/
