+++ lib/README.md
@@ -1,0 +1,4 @@
+albiondata-client/lib
+=====================
+
+This package contains the structs needed to work with our json uploads.
